introduce concrete types with fixed type parameters
this way a consumer of yauaa does not have to deal with type
parameters.
