+++ analyzer/src/main/java/nl/basjes/parse/useragent/UserAgentAnalyzer.java
@@ -18,118 +18,20 @@
 package nl.basjes.parse.useragent;
 
 import com.esotericsoftware.kryo.DefaultSerializer;
-import com.esotericsoftware.kryo.Kryo;
-import org.apache.commons.collections4.map.LRUMap;
 
 import java.io.Serializable;
 
 @DefaultSerializer(UserAgentAnalyzer.KryoSerializer.class)
-public class UserAgentAnalyzer extends UserAgentAnalyzerDirect implements Serializable {
-    public static final int DEFAULT_PARSE_CACHE_SIZE = 10000;
+public final class UserAgentAnalyzer extends AbstractUserAgentAnalyzer implements Serializable {
 
-    private int cacheSize = DEFAULT_PARSE_CACHE_SIZE;
-    private LRUMap<String, UserAgent> parseCache = null;
-
-    protected UserAgentAnalyzer() {
-        super();
-        initializeCache();
+    public static UserAgentAnalyzerBuilder  newBuilder() {
+        return new UserAgentAnalyzerBuilder(new UserAgentAnalyzer());
     }
 
-    public static class KryoSerializer extends UserAgentAnalyzerDirect.KryoSerializer {
-        public KryoSerializer(Kryo kryo, Class<?> type) {
-            super(kryo, type);
-        }
-    }
+    public static final class UserAgentAnalyzerBuilder extends AbstractUserAgentAnalyzerBuilder<UserAgentAnalyzer, UserAgentAnalyzerBuilder>{
 
-    public void disableCaching() {
-        setCacheSize(0);
-    }
-
-    /**
-     * Sets the new size of the parsing cache.
-     * Note that this will also wipe the existing cache.
-     *
-     * @param newCacheSize The size of the new LRU cache. As size of 0 will disable caching.
-     */
-    public void setCacheSize(int newCacheSize) {
-        cacheSize = Math.max(newCacheSize, 0);
-        initializeCache();
-    }
-
-    private void initializeCache() {
-        if (cacheSize >= 1) {
-            parseCache = new LRUMap<>(cacheSize);
-        } else {
-            parseCache = null;
-        }
-    }
-
-    public int getCacheSize() {
-        return cacheSize;
-    }
-
-    @Override
-    public synchronized UserAgent parse(UserAgent userAgent) {
-        if (userAgent == null) {
-            return null;
-        }
-        userAgent.reset();
-
-        if (parseCache == null) {
-            return super.parse(userAgent);
-        }
-
-        String userAgentString = userAgent.getUserAgentString();
-        UserAgent cachedValue = parseCache.get(userAgentString);
-        if (cachedValue != null) {
-            userAgent.clone(cachedValue);
-        } else {
-            cachedValue = new UserAgent(super.parse(userAgent));
-            parseCache.put(userAgentString, cachedValue);
-        }
-        // We have our answer.
-        return userAgent;
-    }
-
-    public static UserAgentAnalyzerBuilder<? extends UserAgentAnalyzer, ? extends UserAgentAnalyzerBuilder<?, ?>> newBuilder() {
-        return new UserAgentAnalyzerBuilder<>(new UserAgentAnalyzer());
-    }
-
-    @SuppressWarnings("unchecked")
-    public static class UserAgentAnalyzerBuilder<UAA extends UserAgentAnalyzer, B extends UserAgentAnalyzerBuilder<UAA, B>>
-            extends UserAgentAnalyzerDirectBuilder<UAA, B> {
-        private final UAA uaa;
-
-        public UserAgentAnalyzerBuilder(UAA newUaa) {
+        private UserAgentAnalyzerBuilder(UserAgentAnalyzer newUaa) {
             super(newUaa);
-            this.uaa = newUaa;
-        }
-
-        /**
-         * Specify a new cache size (0 = disable caching).
-         * @param newCacheSize The new cache size value
-         * @return the current Builder instance.
-         */
-        public B withCache(int newCacheSize) {
-            failIfAlreadyBuilt();
-            uaa.setCacheSize(newCacheSize);
-            return (B)this;
-        }
-
-        /**
-         * Disable caching.
-         * @return the current Builder instance.
-         */
-        public B withoutCache() {
-            failIfAlreadyBuilt();
-            uaa.setCacheSize(0);
-            return (B)this;
-        }
-
-        @SuppressWarnings("EmptyMethod") // We must override the method because of the generic return value.
-        @Override
-        public UAA build() {
-            return super.build();
         }
     }
 }
