+++ analyzer/src/test/java/nl/basjes/parse/useragent/serialization/AbstractSerializationTest.java
@@ -55,7 +55,7 @@ public abstract class AbstractSerializationTest {
         LOG.info("Create");
         LOG.info("--------------------------------------------------------------");
 
-        UserAgentAnalyzerTesterBuilder<?, ?> uaab = UserAgentAnalyzerTester
+        UserAgentAnalyzerTesterBuilder uaab = UserAgentAnalyzerTester
             .newBuilder()
             .keepTests()
             .dropDefaultResources()
